# Performance: Adds use of ValueStopwatch instead of Stopwatch