--- a/Microsoft.Azure.Cosmos.Encryption/src/EncryptionDiagnosticsContext.cs
+++ b/Microsoft.Azure.Cosmos.Encryption/src/EncryptionDiagnosticsContext.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
     internal sealed class EncryptionDiagnosticsContext
     {
         private DateTime startTime;
-        private ValueStopwatch stopwatch;
+        private Stopwatch stopwatch;
         private bool isDecryptionOperation;
 
         public EncryptionDiagnosticsContext()
@@ -30,7 +30,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
         public void Begin(string operation)
         {
-            this.stopwatch = ValueStopwatch.StartNew();
+            this.stopwatch = Stopwatch.StartNew();
             this.startTime = DateTime.UtcNow;
 
             switch (operation)