--- a/Microsoft.Azure.Cosmos/src/Metrics.cs
+++ b/Microsoft.Azure.Cosmos/src/Metrics.cs
@@ -10,7 +10,7 @@ namespace Microsoft.Azure.Cosmos
 
     internal sealed class Metrics
     {
-        private readonly ValueStopwatch stopwatch;
+        private ValueStopwatch stopwatch;
 
         public Metrics()
         {