--- a/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/HistogramExtensions.cs
+++ b/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/HistogramExtensions.cs
@@ -211,9 +211,9 @@ namespace HdrHistogram
         /// </remarks>
         public static void Record(this IRecorder recorder, Action action)
         {
-            var start = Stopwatch.GetTimestamp();
+            var start = ValueStopwatch.GetTimestamp();
             action();
-            var elapsed = Stopwatch.GetTimestamp() - start;
+            var elapsed = ValueStopwatch.GetTimestamp() - start;
             recorder.RecordValue(elapsed);
         }
 
@@ -251,12 +251,12 @@ namespace HdrHistogram
             public Timer(IRecorder recorder)
             {
                 _recorder = recorder;
-                _start = Stopwatch.GetTimestamp();
+                _start = ValueStopwatch.GetTimestamp();
             }
 
             public void Dispose()
             {
-                var elapsed = Stopwatch.GetTimestamp() - _start;
+                var elapsed = ValueStopwatch.GetTimestamp() - _start;
                 _recorder.RecordValue(elapsed);
             }
         }