--- a/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/HistogramExtensions.cs
+++ b/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/HistogramExtensions.cs
@@ -23,6 +23,7 @@ using System.IO;
 using System.Linq;
 using HdrHistogram.Iteration;
 using HdrHistogram.Output;
+using Microsoft.Azure.Documents;
 
 namespace HdrHistogram
 {