--- a/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/OutputScalingFactor.cs
+++ b/Microsoft.Azure.Cosmos/src/OSS/HdrHistogram/OutputScalingFactor.cs
@@ -17,6 +17,7 @@
  */
 
 using System.Diagnostics;
+using Microsoft.Azure.Documents;
 
 namespace HdrHistogram
 {