--- a/Microsoft.Azure.Cosmos/src/Query/Core/Metrics/SchedulingStopwatch.cs
+++ b/Microsoft.Azure.Cosmos/src/Query/Core/Metrics/SchedulingStopwatch.cs
@@ -16,17 +16,17 @@ namespace Microsoft.Azure.Cosmos.Query.Core.Metrics
         /// <summary>
         /// Stopwatch used to measure turnaround time.
         /// </summary>
-        private readonly Stopwatch turnaroundTimeStopwatch;
+        private readonly ValueStopwatch turnaroundTimeStopwatch;
 
         /// <summary>
         /// Stopwatch used to measure response time.
         /// </summary>
-        private readonly Stopwatch responseTimeStopwatch;
+        private readonly ValueStopwatch responseTimeStopwatch;
 
         /// <summary>
         /// Stopwatch used to measure runtime.
         /// </summary>
-        private readonly Stopwatch runTimeStopwatch;
+        private readonly ValueStopwatch runTimeStopwatch;
 
         /// <summary>
         /// Number of times the process was preempted.
@@ -43,9 +43,9 @@ namespace Microsoft.Azure.Cosmos.Query.Core.Metrics
         /// </summary>
         public SchedulingStopwatch()
         {
-            this.turnaroundTimeStopwatch = new Stopwatch();
-            this.responseTimeStopwatch = new Stopwatch();
-            this.runTimeStopwatch = new Stopwatch();
+            this.turnaroundTimeStopwatch = new ValueStopwatch();
+            this.responseTimeStopwatch = new ValueStopwatch();
+            this.runTimeStopwatch = new ValueStopwatch();
         }
 
         /// <summary>