--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/ClientTests.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/ClientTests.cs
@@ -132,8 +132,8 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 {
                     tasks.Add(this.ReadNotFound(container));
                 }
-            
-                Stopwatch sw = Stopwatch.StartNew();
+
+                ValueStopwatch sw = ValueStopwatch.StartNew();
                 while(this.TaskStartedCount < 10 && sw.Elapsed.TotalSeconds < 2)
                 {
                     await Task.Delay(TimeSpan.FromMilliseconds(50));