--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/QueryPerfTest/ContentSerializationPerformanceTests.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/QueryPerfTest/ContentSerializationPerformanceTests.cs
@@ -112,8 +112,8 @@
         private async Task GetIteratorResponse<T>(FeedIterator<T> feedIterator)
         {
             MetricsAccumulator metricsAccumulator = new MetricsAccumulator();
-            ValueStopwatch totalTime = new ValueStopwatch();
-            ValueStopwatch getTraceTime = new ValueStopwatch();
+            Documents.ValueStopwatch totalTime = new Documents.ValueStopwatch();
+            Documents.ValueStopwatch getTraceTime = new Documents.ValueStopwatch();
             string diagnosticDataPath = Path.GetFullPath(DiagnosticsDataFileName);
             while (feedIterator.HasMoreResults)
             {