--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Query/Metrics/Performance.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Query/Metrics/Performance.cs
@@ -33,7 +33,7 @@ namespace Microsoft.Azure.Cosmos.Performance.Tests.Query.Metrics
         [Benchmark]
         public void TestParse()
         {
-            Stopwatch stopwatch = Stopwatch.StartNew();
+            ValueStopwatch stopwatch = ValueStopwatch.StartNew();
             for (int i = 0; i < 100000; i++)
             {
                 BackendMetricsParser.TryParse(delimitedString, out BackendMetrics backendMetrics);