--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/AsyncCacheTest.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/AsyncCacheTest.cs
@@ -298,7 +298,7 @@ namespace Microsoft.Azure.Cosmos.Tests
         public async Task TestAsyncDeadlock()
         {
             AsyncCache<int, int> cache = new AsyncCache<int, int>();
-            Stopwatch stopwatch = new Stopwatch();
+            ValueStopwatch stopwatch = new ValueStopwatch();
 
             stopwatch.Start();
             await Task.Factory.StartNew(() =>