--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/GlobalEndpointManagerTest.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/GlobalEndpointManagerTest.cs
@@ -320,7 +320,7 @@ namespace Microsoft.Azure.Cosmos
                 slowPrimaryRegionHelper.ShouldDelayRequest = (uri) => uri == defaultEndpoint || uri == new Uri(databaseAccount.ReadLocationsInternal.First().Endpoint);
                 slowPrimaryRegionHelper.ShouldFailRequest = slowPrimaryRegionHelper.ShouldDelayRequest;
 
-                Stopwatch stopwatch = Stopwatch.StartNew();
+                ValueStopwatch stopwatch = ValueStopwatch.StartNew();
                 globalEndpointResult = await GlobalEndpointManager.GetDatabaseAccountFromAnyLocationsAsync(
                     defaultEndpoint: defaultEndpoint,
                     locations: new List<string>(){
@@ -548,7 +548,7 @@ namespace Microsoft.Azure.Cosmos
             DefaultTrace.TraceSource.Listeners.Add(new TestTraceListener { Callback = TraceHandler });
             DefaultTrace.InitEventListener();
 
-            Stopwatch stopwatch = Stopwatch.StartNew();
+            ValueStopwatch stopwatch = ValueStopwatch.StartNew();
             // Wait for the trace message saying the background refresh occurred
             while (!isGlobalEndpointRefreshStarted)
             {