# Pull Request Template

## Description
This PR is for a performance improvement to use a new ValueStopwatch type instead of the Stopwatch type, which reduces allocations. 

Micro Benchmark Test Results:
|                     Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|--------------------------- |----------:|----------:|----------:|-------:|----------:|
|        StopwatchAllocation | 3.8487 ns | 0.1114 ns | 0.2876 ns | 0.0093 |      40 B |
|   ValueStopwatchAllocation | 0.1195 ns | 0.0400 ns | 0.1180 ns |      - |         - |
|           StopwatchAsField | 8.4661 ns | 0.1948 ns | 0.2600 ns | 0.0148 |      64 B |
| ValueStopwatchBoxedAsField | 1.8442 ns | 0.0438 ns | 0.1272 ns |      - |         - |

## Type of change
- [x] New feature (non-breaking change which adds functionality)

## Closing issues
Part 2 of 2, closes #3274