@imanvt 2022-07-05
issues were caused by the stopwatch being readonly, fixed and reopening PR. Ready for review